backend/requirements.txt
Django>=5.2.7
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
mysqlclient>=2.2.0
django-cors-headers>=4.3.0



